fixed a problem with a get Method

+++ graphr/src/graphr/data/PrimData.java
@@ -84,7 +84,7 @@ public class PrimData implements JsonReadableWritable {
 	//-- Getters
 	
 	public int i() {
-		return ((Long) o).intValue();
+		return ((Integer) o).intValue();
 	}
 
 	public long l() {
